Changed the way the dynamic officer updater works. Fixed bugs associated with it. Fully functional now, without any GUI graphical issues.
When screen is initiated it will now build a widget for all of the officers, then store them in an array. When an officer goes online, it uses this widget instead of creating a new one. Creating widgets after screen initializing creates a bug where the widget is broken and doesn't do anything.